add folder to Go
